refactoring scoreboard app, scoreboard is now container component, refactored stopwatch component
