+++ package.json
@@ -26,7 +26,9 @@
   "dependencies": {
     "react": "^15.4.2",
     "react-dom": "^15.4.2",
+    "react-redux": "^5.0.2",
     "react-router": "^3.0.2",
+    "redux": "^3.6.0",
     "style-loader": "^0.13.1",
     "url-loader": "^0.5.7"
   },
