+++ src/index.ejs
@@ -22,6 +22,6 @@
   <title>Title</title>
 </head>
 <body>
-<div id="app"></div>
+<div id="root"></div>
 </body>
 </html>
